runes: change Rune() constructor to make more sense.
In our tests, and when demarshalling, we generate runes with an
already-known authcode.  But in the normal cases, we want a the
rune constructor to build on an existing authcode.

And we want to *strongly* encourage adding a unique_id, so make that
the first argument.  An assertion catches old code using it the old
order.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>